Added modqueue chan and auto pokenav badge assigns on mod react
A configurable list of channels can bet set for the bot to listen on.
A configureable set of emoji:badge id pairs can be set
When a mod reacts to a message in a listened channel, the bot will grant the associated badge to the user who posted the message

Configurable modqueue channel added. if enabled, bot will prompt mod whether to grant badge before doing so